# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API Watch Store",
        "description": "Danh sách API cho ứng dụng Khách hàng (ReactJS)",
        "contact": {
            "email": "admin@watchstore.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Khách hàng - Sản phẩm"
                ],
                "summary": "Lấy danh sách sản phẩm (Tìm kiếm, Lọc, Phân trang)",
                "operationId": "88ff1c4110e2af09b78108755b5fa54f",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Tìm theo tên",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "Mã thương hiệu (VD: TH01)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Mã loại (VD: LO01)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Giá tối thiểu",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Giá tối đa",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sắp xếp (price_asc, price_desc, latest)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Khách hàng - Sản phẩm"
                ],
                "summary": "Chi tiết sản phẩm",
                "operationId": "d83ce5b842add284e9536019968dcb83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin chi tiết"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
        "/products/{id}/related": {
            "get": {
                "tags": [
                    "Khách hàng - Sản phẩm"
                ],
                "summary": "Sản phẩm liên quan (Cùng thương hiệu hoặc cùng loại)",
                "operationId": "3670ae265a73e61929bbd1c928babd08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách 4 sản phẩm liên quan"
                    }
                }
            }
        },
...
        "/products/{id}/review": {
            "post": {
                "tags": [
                    "Khách hàng - Sản phẩm"
                ],
                "summary": "Gửi đánh giá sản phẩm (Yêu cầu đăng nhập)",
                "operationId": "3a13fa94b0cdad5522a7ff10a81f33ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Sản phẩm rất tốt!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}